better implementation of subsref for Point (returns a double if indexed as, e.g. pt(2:3)) resolveConstraints consumes and returns Points instead of doubles
git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@5248 c9849af7-e679-4ec6-a44e-fc146a885bd3

+++ examples/RimlessWheel/test/testLCPgradients.m
@@ -4,7 +4,7 @@ options.floating = true;
 m = PlanarRigidBodyManipulator('../RimlessWheel.urdf',options);
 
 p = TimeSteppingRigidBodyManipulator(m,.01);
-x0 = p.resolveConstraints([0;0;randn;5*rand;randn;5*rand]);
+x0 = double(p.resolveConstraints([0;0;randn;5*rand;randn;5*rand]));
 v = p.constructVisualizer();
 v.draw(0,x0);
